Add optional promise-aware behavior to frost-modal-form
 ### Overview

This adds optional flags and behavior to the `frost-modal-form`
component which allows it to behave in a convenient way when `onConfirm`
returns a promise.

 #### Disabling the confirm button until `onConfirm` resolves

When `onConfirm` triggers an action that is either expensive or not
idempotent, we now can prevent spamming of the Confirm button by setting
the `closeOnConfirm` to false, and `disableConfirmUntilOnConfirmResolves`
to true.

The text to be shown while disabled can be changed by overriding the
`disabledConfirm` property.

```
{{frost-modal-form
  closeOnConfirm=false
  disableConfirmUntilOnConfirmResolves=true
  // optional override of button props:
  disabledConfirm=(object
    disabled=true
    isVisible=  // (true)
    text= // ('Confirm')
  )
  ....
}}
```

 #### Automatically close once `onConfirm` resolves
```
{{frost-modal-form
  closeOnConfirm=false
  closeAfterOnConfirmResolves=true
  ....
}}
```

 #### Demo updates

The modal form demo has been expanded with an extra
section (`/#/form-async`) demonstrating the new behavior.

 # Changelog
- Add optional `closeAfterOnConfirmResolves` flag
- Add optional `disableConfirmUntilOnConfirmResolves` flag
- Allow forcing disabled confirm button state until resolve
- Show demo notifications above the blur layer
- Add support for title tooltips on modal buttons
- Update to ember-test-utils@^8.1.0
- Add eslint rules to allow existing test names
- Fix linting per updated linting rules
- Add chai-jquery, sinon-chai devDependencies